requirements_local.txt
aiohttp==3.8.3
charset-normalizer>=2.0,<3.0
playwright
docx2txt
python-pptx
Pyarrow
unstructured[all-docs]
poppler-utils
python-magic
magika
pandas
colorama
requests
pillow
pytesseract
cssutils
beautifulsoup4
torch
llmlingua
PyMuPDF
pdf2image
python-magic
nltk
scikit-learn